pyproject.toml
[project]
name = "fii-orchestrator"
version = "0.1.0"
description = ""
authors = [
    {name = "Vitoria Sousa",email = "vitoria.sousa@stone.com.br"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "feedparser (>=6.0.11,<7.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "polars (>=1.32.2,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "unidecode (>=1.4.0,<2.0.0)",
    "yfinance (>=0.2.65,<0.3.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "requests (>=2.31.0,<3.0.0)",
    "numpy (>=1.24.0,<2.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "duckdb (>=1.3.2,<2.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "pydantic-settings (>=2.0.0,<3.0.0)",
    "dependency-injector (>=4.41.0,<5.0.0)",
    "asyncpg (>=0.29.0,<0.30.0)",
    "sqlalchemy[asyncio] (>=2.0.0,<3.0.0)",
    "alembic (>=1.13.0,<2.0.0)",
    "clickhouse-connect (>=0.7.0,<0.8.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.8"
black = "^25.1.0"
pre-commit = "^4.3.0"
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.11.0"
httpx = "^0.28.1"

[tool.poetry]
name = "fii-orchestrator"
version = "0.1.0"
description = "Orquestrador de IA multiagente para carteira de FIIs"
readme = "README.md"
# IMPORTANTE: mapear o pacote para a raiz src/
packages = [{ include = "fii_orchestrator", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
# ...demais dependências
